Revert "Split AbstractArrays and ConcreteArrays classes into files (#132)"
This reverts commit 2c5f84dd4a8fe5c7dc40e40783fc99dd75a7c04f.
