+++ src/scene/Mesh.hpp
@@ -26,7 +26,6 @@
 #include <scene/ASBuildScratchpad.hpp>
 
 #include <filesystem>
-#include <memory/DeviceSyncArray.hpp>
 
 
 struct Mesh : APIObject<Mesh>
