+++ test/src/ArrayTest.cpp
@@ -1,9 +1,10 @@
 #include <gtest/gtest.h>
 #include <gmock/gmock-matchers.h>
 
+#include <memory/AbstractArrays.hpp>
+#include <memory/ConcreteArrays.hpp>
 #include <memory/InvalidArrayCast.hpp>
 #include <memory/DeviceAsyncArrayManager.hpp>
-#include <memory/HostPinnedArray.hpp>
 
 using namespace ::testing;
